Increasing the stability of the end-to-end tester utility
If, for any reason (for example: the computer is overloaded), the keyup
event cannot be sent quikly enough after the keydown event, it may happen
that the keydown event is received multiple times, which was previously
making the test fail. This commit makes the test a bit less strict and
allows multiple keydown events before the keyup event.